🧼 Clean up utils
